Fixed abstract-inclusive registered models listing with app specified.
The various get_models methods of Django don't give an option to include
abstract models, so a somewhat different solution to filter models by app
was needed. This had not been a problem for management commands as they
don't deal in abstract models, nevertheless should be fixed for other use
cases.