Made test_registration only count models from modeltranslation.tests app.
The only thing that was preventing the test from failing in the context
of a project with its own models registered for translation was the
translator registry duplication, which was removed in 1e566f4e.

The models count change is due to no longer considering the single
registered model from "managed_app".